requirements.txt
aws_xray_sdk==2.14.0
channels==4.3.1
Django==5.2.5
django-storages==1.14.6
djangorestframework==3.16.1
djoser==2.3.3
drf_nested_routers==0.94.2
drf_yasg==1.21.10
psycopg2==2.9.10
requests==2.32.4
rest_framework_simplejwt==0.0.2
urllib3<1.26
boto3
psycopg2
djoser
pytest
django-filter
Pillow

